src/pq_crypto.rs
// src/pq_crypto.rs

// Placeholder for post-quantum cryptographic library imports
// In a real implementation, you would use actual libraries like pqcrypto or similar

pub enum PQAlgorithm {
    Dilithium,
    Falcon,
}

pub fn generate_post_quantum_keypair() -> (Vec<u8>, Vec<u8>) {
    // Placeholder for keypair generation
    // Replace with actual keypair generation logic
    (vec![0; 32], vec![0; 64])
}

pub fn sign_message(private_key: &[u8], message: &[u8], _nonce: &[u8]) -> Vec<u8> {
    // Placeholder for signing logic
    // Replace with actual signing using PQ algorithms
    message.to_vec() // For simplicity, return the message as signature
}

pub fn verify_signature(signatures: &[Vec<u8>]) -> bool {
    // Placeholder for signature verification logic
    // Replace with actual verification
    !signatures.is_empty()
}
